+++ controllers/externalCollabs.controller.js
@@ -1,29 +1,21 @@
-const Supplies = require("../models/externalCollabs.model");
+const externalCollabs = require("../models/externalCollabs.model");
 
 exports.getExternalCollabs = async (request, response) => {
     try {
-        const externalCollabs = await externalCollabs.fetchAll();
-        response.status(200).json(externalCollabs);
+        const externalCollaborators = await externalCollabs.fetchAll();
+        response.status(200).json(externalCollaborators);
     } catch (error) {
-        console.error("Error fetching supplies:", error);
-        response.status(500).json({ message: "Failed to fetch supplies." });
+        console.error("Error fetching external collaborators:", error);
+        response.status(500).json({ message: "Failed to fetch external collaborators." });
     }
 };
 
-// Register a external collab
-exports.postExternalCollabs = (request, response) => {
-  const new_ExternalColab = new Usuario(
-    request.body.nombre,
-    request.body.apellidoPaterno,
-    request.body.apellidoMaterno,
-    request.body.fechaNacimiento,
-    request.body.carrera,
-    request.body.correoElectronico,
-    request.body.telefono,
-    request.body.estatus,
-    request.body.reglamentoInterno,
-    request.body.copiaINE,
-    request.body.avisoConfidencialidad,
-    request.body.foto
-  );
-}
+exports.getExternalCollabsByID = async (request, response) => {
+  try {
+    const externalCollabId = await externalCollabs.fetchOneByID(request.params.id);
+    response.status(200).json(externalCollabId);
+  } catch (error) {
+    console.error("Error fetching external collab by ID:", error);
+    response.status(500).json({ message: "Failed to fetch external collab by ID." });
+  }
+};