+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [0.1.0] - 01/11/2025  
 
 ### Added
+ - Sequence diagram for User Story of viewing external collaborator and all of them.
+ - Logic to have the routes to view external collaborators (all and by ID).
  - Modified supplies model to correctly fetch data from the database (search, order, filter)
  - Updated supplies controller to handle errors when fetching supplies by ID.
  - Method to handle post request in index.js.